Fazendo melhorias no menu da navbar na versão mobile com uma modal.
