+++ src/components/MenuElements.module.css
@@ -1,16 +1,23 @@
 .activeContainer {
+    display: flex;
+    align-items: center;
+
     background-color: var(--yellow-200);
-    padding: 5px 10px;
+    height: 55px;
+    margin: -10px 0;
 }
 
 .activeContainer a {
+    font-size: 22px;
     color: var(--red);
+    padding-left: 20px;
 }
 
 a {
-    font-size: 18px;
+    font-size: 22px;
     color: var(--white);
     text-decoration: none;
+    padding-left: 20px;
 }
 
 a:hover {
