+++ src/components/NavbarMobile/index.jsx
@@ -1,37 +1,20 @@
 import styles from "./NavbarMobile.module.css"
 import { useState } from "react";
 
-import { MenuElements } from "../MenuElements";
-import { SearchBar } from "../SearchBar";
+import { NavbarModal } from "../NavbarModal";
 
 export function NavbarMobile() {
     const [showMenu, setShowMenu] = useState(false);
-    const menuOptions = [
-        {optionName: "Home", optionUrl: "/"}, {optionName: "PodCast", optionUrl: "/PodCast"}, {optionName: "Cinema", optionUrl: "#"},
-        {optionName: "Livros", optionUrl: "#"}, {optionName: "Musícas", optionUrl: "#"}, {optionName: "Jogos", optionUrl: "#"}
-    ];
-
-    const toggleMenu = () => {
-        setShowMenu(!showMenu)
-    }
-
+    
     return (
         <>
-            <nav className={`${styles.menuSandwich} ${ showMenu ? styles.show : '' }`}>
-                {menuOptions.map(function(data) {
-                    return (
-                        <MenuElements name={data.optionName} screenUrl={data.optionUrl} />
-                    )
-                })}
-
-                <SearchBar />
-            </nav>
-
-            <div className={styles.menuButton} onClick={toggleMenu}>
+            <div className={styles.menuButton} onClick={() => setShowMenu(true)}>
                 <span className={styles.linha}></span>
                 <span className={styles.linha}></span>
                 <span className={styles.linha}></span>
             </div>
+
+            <NavbarModal showMenu={showMenu} setShowMenu={() => setShowMenu(!showMenu)} />
         </>
     );
 }
