+++ src/components/NavbarModal/NavbarModal.module.css
@@ -1,0 +1,38 @@
+.bg {
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    background-color: var(--gray-fill-400);
+    z-index: 1000;
+    
+    display: flex;
+    justify-content: flex-end;
+}
+
+.modalContainer {
+    position: fixed;
+    height: 100vh;
+    width: 85vw;
+    background-color: var(--black);
+}
+
+.iconPosition {
+    display: flex;
+    justify-content: flex-end;
+}
+
+.icon {
+    margin-top: 50px;
+    margin-right: 25px;
+    font-size: 35px;
+}
+
+.menuContainer {
+    display: flex;
+    justify-content: flex-start;
+    flex-direction: column;
+    gap: 30px;
+    margin-top: 50px;
+}
